refactor(core): split bridge/handlers.py into domain modules
Split the monolithic 1,471-line handlers.py file into domain-specific
modules for improved maintainability, testability, and code ownership.

Changes:
- Extract shared.py with base state management and library accessors
- Extract lifecycle.py for process lifecycle handlers
- Extract debugger.py for breakpoint and stepping handlers
- Extract codegen.py for code generation handlers
- Extract webui_spy.py for web UI inspection handlers
- Extract desktopui_spy.py for desktop UI inspection handlers
- Maintain backward compatibility via __init__.py facade

All 31 handler methods preserved with identical signatures.
All 381 core tests pass.
